Trim comments to the terse bar; hoist a shared firmware-job test factory
Cut the over-explained comments and module docstrings added in this PR down to
the non-obvious why. Extract test/_make-firmware-job.ts (makeFirmwareJob) and
use it from the install-chain and jobs-dedup tests instead of each carrying a
full FirmwareJob literal.
