+++ test/components/app-shell-jobs-dedup.test.ts
@@ -1,48 +1,10 @@
-/**
- * Terminal-job dedup in the live jobs context keys on (config, type).
- *
- * An install is a COMPILE then a dependent UPLOAD sharing a configuration
- * (backend #1131). Dedup-per-config-alone let the upload's completion evict
- * the compile from the live list (the build log vanished until a reload).
- * Keying on type too keeps both, matching the backend's retention.
- */
+// Live-jobs dedup keys on (config, type) so an install's compile and upload
+// both survive — the upload's terminal must not evict the compile (#1131).
 import { describe, expect, it } from "vitest";
-import {
-  type FirmwareJob,
-  JobSource,
-  JobStatus,
-  JobType,
-} from "../../src/api/types/firmware-jobs.js";
+import { JobStatus, JobType } from "../../src/api/types/firmware-jobs.js";
 import type { ESPHomeApp } from "../../src/components/app-shell.js";
 import { handleJobEvent } from "../../src/components/app-shell/jobs.js";
-
-function makeJob(overrides: Partial<FirmwareJob> = {}): FirmwareJob {
-  return {
-    job_id: "job-1",
-    configuration: "kitchen.yaml",
-    job_type: JobType.COMPILE,
-    status: JobStatus.COMPLETED,
-    created_at: "2026-01-01T00:00:00Z",
-    started_at: null,
-    completed_at: null,
-    exit_code: null,
-    output: [],
-    error: null,
-    port: "OTA",
-    new_name: "",
-    depends_on: "",
-    progress: null,
-    source: JobSource.LOCAL,
-    source_pin_sha256: "",
-    source_label: "",
-    source_esphome_version: "",
-    remote_peer: "",
-    remote_peer_label: "",
-    device_name: "",
-    device_friendly_name: "",
-    ...overrides,
-  };
-}
+import { makeFirmwareJob as makeJob } from "../_make-firmware-job.js";
 
 function makeHost(): ESPHomeApp {
   return {
@@ -53,18 +15,21 @@ function makeHost(): ESPHomeApp {
   } as unknown as ESPHomeApp;
 }
 
+const done = (overrides: Parameters<typeof makeJob>[0]) =>
+  makeJob({ status: JobStatus.COMPLETED, ...overrides });
+
 describe("live jobs dedup (config, type)", () => {
   it("keeps an install's compile and upload terminals for the same config", () => {
     const host = makeHost();
     handleJobEvent(
       host,
       "job_completed",
-      makeJob({ job_id: "c", job_type: JobType.COMPILE })
+      done({ job_id: "c", job_type: JobType.COMPILE })
     );
     handleJobEvent(
       host,
       "job_completed",
-      makeJob({ job_id: "u", job_type: JobType.UPLOAD, depends_on: "c" })
+      done({ job_id: "u", job_type: JobType.UPLOAD, depends_on: "c" })
     );
     expect(new Set(host._firmwareJobs.keys())).toEqual(new Set(["c", "u"]));
   });
@@ -74,12 +39,12 @@ describe("live jobs dedup (config, type)", () => {
     handleJobEvent(
       host,
       "job_completed",
-      makeJob({ job_id: "c1", job_type: JobType.COMPILE })
+      done({ job_id: "c1", job_type: JobType.COMPILE })
     );
     handleJobEvent(
       host,
       "job_completed",
-      makeJob({ job_id: "c2", job_type: JobType.COMPILE })
+      done({ job_id: "c2", job_type: JobType.COMPILE })
     );
     expect(new Set(host._firmwareJobs.keys())).toEqual(new Set(["c2"]));
   });
